Moved NetworkSession protocol and its implementations to their own dedicated file.
