+++ Sources/FlowMagic-SDK/Services/WebService.swift
@@ -12,48 +12,6 @@ enum NetworkError: Error {
     case invalidServerResponse
 }
 
-private extension URLResponse {
-    var isSuccessful: Bool {
-        guard let httpResponse = self as? HTTPURLResponse else {
-            return false
-        }
-
-        return httpResponse.statusCode == 200
-    }
-}
-
-protocol NetworkSession {
-    func loadData(from url: URL) async throws -> (Data, HTTPURLResponse)
-}
-
-extension URLSession: NetworkSession {
-    func loadData(from url: URL) async throws -> (Data, HTTPURLResponse) {
-        let (data, response) = try await self.data(from: url)
-        if let httpResponse = response as? HTTPURLResponse {
-            return (data, httpResponse)
-        } else {
-            throw NetworkError.invalidServerResponse
-        }
-
-    }
-}
-
-class MockNetworkSession: NetworkSession {
-    var mockData: Data?
-    var mockResponse: HTTPURLResponse?
-    var mockError: Error?
-
-    func loadData(from url: URL) async throws -> (Data, HTTPURLResponse) {
-        if let error = mockError {
-            throw error
-        }
-        guard let data = mockData, let response = mockResponse else {
-            throw NetworkError.invalidServerResponse
-        }
-        return (data, response)
-    }
-}
-
 protocol WebService {
     func loadUrlData(resource: String) async throws -> ScreenFlowModel
 }
